Added the ability to move topics upwards in the document
User are now able to perform movements with topics only in the upwards
direction. Although this is a bit clunky since when moving a field from
a lower depth to a higher one it'll jump straight above the previous
item, the more natual way of doing this would be to first simply indent
the field below the first higher-depth field.