Moving item upwards as child of previous element in GtkTreeView
Implemented the logic that when moving an item upwards it's moved inside
a previous item as the last item when the previous item has child items.
This is to mimic the user's expected behaviour and how it was
implemented in Windows, avoiding jumps when moving items around.

A method to extract and rebuild the field objects from the GtkTreeView
was also created to support this functionality.