Fixed the paths of the eVC PocketPC project port
Went through and fixed all the paths in the pre-Windows Mobile PocketPC
port project, the one that uses eMbedded Visual C++ 3.

Even after a lot of hacking around, the menu bar is still not showing up
in the built application. This is a known issue and I was never able to
fix it in the past.